Moved data holders to data loader class
